remove ahm old ahm info module
