--- novawallet/Modules/AHMInfo/AHMInfoWireframe.swift
@@ -1,10 +1,0 @@
-import Foundation
-
-final class AHMInfoWireframe: AHMInfoWireframeProtocol {
-    func complete(from view: AHMInfoViewProtocol?) {
-        view?.controller.dismiss(
-            animated: true,
-            completion: nil
-        )
-    }
-}
